Implemented the MultiSite object. This is designed to allow specification of reactions involving molecules with multiple reaction sites, in an efficient manner. Using this specification, we can now design solvers in MOOSE to efficiently calculate nasty combinatoric chemistry problems. StochSim and other programs have previously come up with ways of doing this.
The MultiSite object manages a number of Molecule objects, which
now are regarded as sites for chemical changes on the MultiSite
molecule. Each of the subsidiary molecules is connected up in the usual
chemical fashion, so the changes to the modeling logic are minimal.

The current MultiSite computes a very approximate version of the
deterministic reaction outcome.

-- Upi