The python API slightly changed:
element is a function implemented in C. It can take a path string, an
existing element or an ematrix and return an instance of the
appropriate moose builtin class.

ObjId/element is renamed "melement". This should not be directly
needed by the users. Python code that used element() earlier will
remain valid.

to_el becomes redundant with this change.