Makefile: simplify detection of python flags
For the last few years, Python 2 is installed as both
python and python2, so the rules can be simplified. As
defined, INSTALLED_PYTHON was used to define flags which
only make sense on UNIX-y systems, since the paths for
Windows will be completely different. So the whole block
can just be removed.